Correction module card.js, main.js
