+++ js/data.js
@@ -55,8 +55,6 @@ const PHOTOS = [
   'http://o0.github.io/assets/images/tokyo/hotel3.jpg',
 ];
 
-const SIMILAR_AD_AMOUNT = 10;
-
 
 /**
  * Функция генерирования объекта
@@ -70,7 +68,7 @@ function createAd() {  // const createAd = () => {
 
   return {
     author: {
-      avatar: 'img/avatars/users0' + getRandomAvatarIndex(1, 8) + '.png',
+      avatar: 'img/avatars/user0' + getRandomAvatarIndex(1, 8) + '.png',
     },
     offer: {
       title: getRandomElements(HEADERS),
@@ -93,6 +91,6 @@ function createAd() {  // const createAd = () => {
 }
 
 /*const createAds = new Array(SIMILAR_AD_AMOUNT).fill(1).map(() => createAd());*/
-const createAds = () => new Array(SIMILAR_AD_AMOUNT).fill(1).map(() => createAd());
+const createAds = (COUNT) => new Array(COUNT).fill(null).map(() => createAd());
 
-export {HEADERS, TYPES, CHECKTIME, FEATURES, DESCRIPTIONS, PHOTOS, SIMILAR_AD_AMOUNT, createAd, createAds};
+export {HEADERS, TYPES, CHECKTIME, FEATURES, DESCRIPTIONS, PHOTOS, createAd, createAds};
