+++ js/main.js
@@ -1,7 +1,11 @@
 import './data.js';
 import './util.js';
-import './cards.js';
+import './card.js';
 
 import {createAds} from './data.js';
+import {similarListElement, renderCard} from './card.js';
 
-console.log(createAds);
+const COUNT = 10;
+const dataAds = createAds(COUNT);
+
+similarListElement.appendChild(renderCard(dataAds[0]));
